Изменил компоненты react
